+++ src/lockss/debugpanel/__main__.py
@@ -28,6 +28,6 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-import lockss.debugpanel.cli
+from .cli import main
 
-lockss.debugpanel.cli.main()
+main()